requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
